test(TableV2): add test
