+++ packages/core/client/src/schema-component/antd/table-v2/__tests__/table-v2.test.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { render, sleep } from 'testUtils';
+import App1 from '../demos/demo1';
+import App2 from '../demos/demo2';
+
+describe('TableV2', () => {
+  it('basic', async () => {
+    const { container } = render(<App1 />);
+
+    await sleep(300);
+    expect(container).toMatchSnapshot();
+  });
+
+  it('tree table', () => {
+    const { container } = render(<App2 />);
+    expect(container).toMatchSnapshot();
+  });
+});
